do not check the sizes thanks to the mappings
Now that the arguments are not positional inside the lists of keys
and values, the code does not need to check whether the sizes do
match.
The input values are seen as a bank of values that the config file
demands to use. The code then only picks the values it needs to
apply to the hunks of the current config file.