remove the clipping on the hue
Clipping the saturation and the brightness between 0 and 1 does
make sense as these values only leave on that segment. However, the
hue value lies on a clock, i.e. 1 wraps back to 0. Thus the script
should not clip the hue, otherwise everything tends to turn red
when applying a big bias.