add `amtoine-gh-notify` to receive GitHub notifications (#23)
* add a simple script to get notified from GitHub

* make the script executable

* rename the script to be more coherent

* fix the top-level notifications with better info

* use readarray to build the different arrays

* refactor into tool functions to make it faster

Prior to that commit, the script had to pull from the GitHub API
each time an information was required. This is highly inefficient
as all the information is already contained in the first
`gh api notifications --jq '.'`! But it was discard each time...
This commit refactors the script into tool sub-functions and, most
importantly, stores the results of the first `gh api notifications`
inside a temporary file on the filesystem, making the following
pulls much faster!

* add a tool function to get fields from json file

* rename the script to "amtoine"

Co-authored-by: a2n-s <44101798+AntoineStevan@users.noreply.github.com>